Use existing ctx handling instead of early return
This commit removes the unintentional masking of the cancelled
context as the reason for the early return out of the retry
loop. This is accomplished by removing the duplicate context
cancellation check and early return and instead relying on the
existing context cancellation check at the top of the loop.

refs GH-88