Fix context cancellation timeout behavior
Prior support of retries and retries delay flags was
unintentionally limited to a hard-coded 5s context cancellation
value. This commit removes that hard-coded limit and instead
sets an application timeout (context cancellation) value
calculated based on a set multipler of 2 and user-specified
retries (default of 2) and retries delay (default of 2) flag
values. By default, the calculation provides an app timeout
of 8s. If the user specifies retries value of 10 and retries
delay of 2 this results in a 40s timeout.

This commit also adds a WARNING message if the calculated
timeout exceeds the default 30s notification timeout applied
by Nagios. If this ends being annoying, we can change the
behavior and place this WARNING behind an optional (future)
flag intended to toggle WARNING messages on/off as desired.

Since this is the first of such WARNING messages, we emit
the message by default.

refs GH-130