Update branding support to better reflect app role
Previously I believe that the branding text gave the wrong
impression regarding the source of the message. This commit
attempts to make clear that this app is the delivery agent,
not the generator of the message.

If provided by the user, the generator or sending application
is noted in the branding text to help anyone viewing the
message better understand the source of the message.

As before, message branding remains optional.

refs GH-149