Add linting using unstable & oldstable containers
Use the same matrix of containers for linting in the
`Validate Codebase` GHAW that we're already using
for testing and building jobs.

This applies the minimum linting requirements in addition
to testing "unstable" linting options that may become the
new baseline in the future.

One notable difference is that out of the matrix of containers
used for linting we mark the unstable container as "experimental"
and configure the job to ignore linting errors generated by
that container. This effectively makes any linting output from
the unstable container informational only as intended.

refs GH-206