Add initial support for user mentions
Initial support for user mentions is provided
by way of new alpha release of `atc0005/go-teams-notify`
library. Due to limitations of the library, user mentions
are not compatible with specific existing features:

- message title
- URL "buttons"
- message border coloring

Future support for user mentions is expected to provide
most (or perhaps all) of this functionality.

README has been updated to provide an example of the new
user mention functionality.

refs GH-197