Remove automatic display of help text on err
Remove usage information when CLI flag validation fails as the
error output should be the primary focus; displaying help text
and error text together is too noisy to be truly helpful.

If/when needed, the user can use the `-h` or `--help` flags to
be reminded of supported flags and values.

refs GH-532